+++ CHANGELOG.md
@@ -116,3 +116,4 @@
 ## 01.08.2025
 - New greet progs added for NPCs.
 - say, shout and tell are colorized.
+- "all" added to do_sacrifice() and number_range() improved.